feat: enhance comment functionality with like feature;

+++ api/controllers/comment.controllers.js
@@ -1,12 +1,12 @@
-import { errorHandlerMiddleware } from "../middleware/error.middleware.js";
 import Comment from "../models/comment.model.js";
+import { errorHandler } from "../utils/ErrorHandler.js";
 
 export const createComment = async (req, res, next) => {
   try {
     const { postId, userId, comment } = req.body;
 
     if (req.user.id != userId) {
-      return next(errorHandlerMiddleware(401, "Unauthorized access"));
+      return next(errorHandler(401, "Unauthorized access"));
     }
 
     const newComment = new Comment({
@@ -32,7 +32,7 @@ export const getComment = async (req, res, next) => {
       return res.status(200).json({ comments });
     }
 
-    return next(errorHandlerMiddleware(404, "No comments found"));
+    return next(errorHandler(404, "No comments found"));
   } catch (error) {
     next(error);
   }
@@ -41,7 +41,15 @@ export const getComment = async (req, res, next) => {
 export const editComment = async (req, res, next) => {
   const reqComment = req.body.comment;
   const { commentId } = req.params;
-
+  const commentToEdit = await Comment.findById(commentId);
+  if (!commentToEdit) {
+    next(errorHandler(404, "No comment found"));
+  }
+  if (req.user.id !== commentToEdit.userId && req.user.isAdmin) {
+    next(
+      errorHandler(401, "Your do not have authorization to edit this comment")
+    );
+  }
   try {
     const comment = await Comment.findByIdAndUpdate(
       commentId,
@@ -58,14 +66,47 @@ export const editComment = async (req, res, next) => {
 };
 
 export const deleteComment = async (req, res, next) => {
-  const commentId = req.params.commentId;
+  const { commentId } = req.params;
 
+  const commentToDelete = await Comment.findById(commentId);
+  if (req.user.id !== commentToDelete.userId && !req.user.isAdmin) {
+    next(errorHandler(401, "Unauthorized access"));
+  }
   try {
     const comment = await Comment.findByIdAndDelete(commentId);
     if (comment) {
       return res.status(200).json({ message: "Comment deleted successfully" });
     }
     return res.status(404).json({ message: "Comment not found" });
+  } catch (e) {
+    next(e);
+  }
+};
+
+export const likeComment = async (req, res, next) => {
+  const { commentId } = req.params;
+  const userId = req.user.id;
+
+  try {
+    const comment = await Comment.findById(commentId);
+
+    if (!comment) {
+      next(errorHandler(404, "Comment not found"));
+    }
+
+    const isLiked = comment.likedBy.indexOf(userId);
+
+    if (isLiked === -1) {
+      comment.likedBy.push(userId);
+      comment.likeCount++;
+      await comment.save();
+      return res.status(200).json(comment);
+    } else {
+      comment.likedBy.splice(isLiked, 1);
+      comment.likeCount--;
+      await comment.save();
+      return res.status(200).json(comment);
+    }
   } catch (e) {
     next(e);
   }
